implemented a message based mechanism to retrieve message by introduce package envelop. the envelop message as the following
first 4 bytes is the size. it is an 32 bit Integer in Network Order.
follow by a payload.
right after the payload, it has a END_OF_MESSAGE magic number defined in
common.h { 0xBA, 0xDB, 0xEE, 0xFB, 0xAD, 0xF0, 0x0D};
[1,2,3,4][payload] [ 0xBA, 0xDB, 0xEE, 0xFB, 0xAD, 0xF0, 0x0D]

the payload is a avro format so it can be produced by any language.